requirements.txt
aniso8601==9.0.1
astroid==2.5.3
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.6
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
flake8==3.9.2
Flask==1.1.2
Flask-Compress==1.9.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
glob2==0.7
idna==2.10
iniconfig==1.1.1
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
parse==1.19.0
parse-type==0.5.2
Pillow==8.3.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.0.1
pylint==2.7.4
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.2.3
pytest-bdd==4.1.0
pytest-cov==2.11.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
